+++ src/components/vehicles/VehicleDetails.js
@@ -14,7 +14,8 @@ class VehicleDetails extends Component {
             mileage:this.props.vehicle.mileage,
             serviceDate:this.props.vehicle.serviceDate,
             imgUrl:this.props.vehicle.imgUrl,
-            updateVehicle:false
+            updateVehicle:false,
+            bookings:[]
         }
         this.handleDelete = this.handleDelete.bind(this);
         this.handleUpdate = this.handleUpdate.bind(this);
@@ -26,28 +27,63 @@ class VehicleDetails extends Component {
 
         const select = document.querySelectorAll('select');
         M.FormSelect.init(select, {});
+
+        const that = this;
+
+        const token = 'Bearer '+ localStorage.token;
+        const headersInfo = {
+            Authorization:token
+        }
+        const data = {
+            email:localStorage.email
+        }
+        console.log(headersInfo);
+        axios.post("http://localhost:8080/GetBookingList",data,{
+            headers:headersInfo
+        }).then(function(res){
+            console.log(res.data);
+            that.setState({
+                bookings:res.data
+            })
+        }).catch(function(error){
+            console.log(error);
+        })
     }
 
     handleDelete() {
-        //ADD VALIDATION
         const config = {
             headers:{
                 Authorization:'Bearer '+ localStorage.token
             }
         }
-        if (window.confirm("Are you sure you want to delete this vehicle?")) {
-            axios.delete("http://localhost:8080/DeleteVehicle/"+ this.state.id,config)
-            .then(function(res){
-                console.log("Vehicle deleted successfully!");
-                alert("Vehicle deleted successfully!");
-                window.location.reload();
-            }).catch(function(error){
-                console.log("Vehicle delete un-successful!\nError : ",error.response);
-                alert("Vehicle delete un-successful!");
-        })
-          } else {
-            alert("Vehicle deletion cancelled");
-          }             
+
+        let canDelete = true;
+        for(let a=0; a<this.state.bookings.length; a++){
+            if(this.state.bookings[a].bookingState==='PickedUp' || this.state.bookings[a].bookingState==='Pending'){
+                if(this.state.bookings[a].vehicle.id === this.state.id){
+                    canDelete=false;
+                }
+            }
+        }
+
+        if(canDelete){
+            if (window.confirm("Are you sure you want to delete this vehicle?")) {
+                axios.delete("http://localhost:8080/DeleteVehicle/"+ this.state.id,config)
+                .then(function(res){
+                    console.log("Vehicle deleted successfully!");
+                    alert("Vehicle deleted successfully!");
+                    window.location.reload();
+                }).catch(function(error){
+                    console.log("Vehicle delete un-successful!\nError : ",error.response);
+                    alert("Vehicle delete un-successful!");
+            })
+              } else {
+                alert("Vehicle deletion cancelled");
+              } 
+        }else{
+            alert("This Vehicle Cannot be Deleted due to existing bookings that include this vehicle");
+        }
+                    
     }
 
     handleUpdate() {
@@ -73,29 +109,21 @@ class VehicleDetails extends Component {
                     <button class="waves-effect waves-light btn-small red lighten-2" onClick={this.handleDelete}>Delete</button>
                 </td>
                 <td class="teal lighten-4 center">
-                    {this.props.vehicle.model}<br/><br/>
-                    <textarea style={{width:200+"px",height:125+"px"}} type="text" placeholder={this.state.description} id="description" onChange={this.handleChange} />
-
+                    <b>{this.props.vehicle.model}</b><br/><br/>
+                    {this.state.description}
                 </td>
                 <td class="center">{this.props.vehicle.category}</td>
                 <td class="teal lighten-4 center">
-                    <input class="center" type="text" placeholder={this.state.rates} id="rates" onChange={this.handleChange} style={{width: 50+"px",height:25+"px"}}/> Euros
+                {this.state.rates} Euros
                 </td>
                 <td>
                     <img class="responsive-img" src={this.state.imgUrl} alt=""/><br/>
-                    <input type="text" placeholder="New Image URL" id="imgUrl" onChange={this.handleChange} style={{height:25+"px"}}/>
                 </td>
                 
                 <td class="teal lighten-4"><b>Fuel Type: </b>{this.props.vehicle.fuelType}
-                <br/><b>Mileage: </b><input type="text" placeholder={this.state.mileage} id="mileage" onChange={this.handleChange} style={{width: 50+"px",height:25+"px"}}/>
+                <br/><b>Mileage: </b>{this.state.mileage}
                 <br/><b>Service Date: </b>
-                <input 
-                    type="text" 
-                    class="datepicker" 
-                    placeholder={this.state.serviceDate!==null?(this.state.serviceDate.toString().split('T')[0]):("")} 
-                    id="serviceDate" 
-                    style={{width: 100+"px",height:25+"px"}}
-                />
+                {this.state.serviceDate!==null?(this.state.serviceDate.toString().split('T')[0]):("")} 
                 <br/><b>Availability: </b>{this.props.vehicle.availability===true?"Available":"Unavailable"}
                 <br/><b>License No: </b>{this.props.vehicle.licenseNo}</td>
